Moved all the Tests
