+++ Core/ALife.Tests/Core/Utility/TestIOHelpers.cs
@@ -1,0 +1,11 @@
+﻿namespace ALife.Tests.Core.Utility
+{
+    /// <summary>
+    /// Tests for the IOHelpers class.
+    /// TODO: implement tests in a preferably platform-independent way.
+    /// </summary>
+    [TestClass]
+    public class TestIOHelpers
+    {
+    }
+}
