--- Core/ALife.Tests/Utility/TestIOHelpers.cs
@@ -1,11 +1,0 @@
-﻿namespace ALife.Tests.Utility
-{
-    /// <summary>
-    /// Tests for the IOHelpers class.
-    /// TODO: implement tests in a preferably platform-independent way.
-    /// </summary>
-    [TestClass]
-    public class TestIOHelpers
-    {
-    }
-}
